Lucia
Working prototype. I'll add the schematic in a few days. It uses an active buzzer, HC-SR04 sensor, infrared receiver and a specific remote. The infrared reader currently allows for turning the cane on and finding it with a repetitive beeping.

Priority is making it programmable to any emitter (use with any remote, much cheaper), then making sound less annoying and adding both female aux and corresponding audio signal; it's automatically detected so as to be optional. Eventually must fix time delay bug that gets the Arduino stuck at the start of processes. Their is also a bug that makes the ranger (names bat() function) get stuck in a loop at maximum frequency/minimum distance (~2 cm) when too close to obstacle.

Ideas: Voltage sensor on palm to detect holding. Button on handle so the remote is only for finding it (could make offer small, single-button remote). Reset button on remote or handle in case of non-activation.